frontend/backend/requirements.txt
Flask==3.1.2
Flask-SocketIO==5.6.0
openai-whisper
numpy
soundfile
reportlab
python-dotenv
pydub
ffmpeg-python
imageio-ffmpeg